Added Certificates route and made name conventions

+++ routes/internship.py
@@ -6,7 +6,7 @@ from dependencies import get_db
 from pydantic import BaseModel 
 
 class InternData(BaseModel):
-    companyname:str
+    company_name:str
     role:str
     description:str
     duration:str
@@ -14,12 +14,12 @@ class InternData(BaseModel):
 
 router = APIRouter(
     prefix="/internship",
-    tags=["internship"]
+    tags=["Internship"]
 )
 
 # to get all the internship data
 @router.get("/")
-def get_internshipdata(db: Session = Depends(get_db)):
+def get_internship_data(db: Session = Depends(get_db)):
     query = select(Internship)
     internship = db.scalars(query).all()
     return {"internship":internship}
@@ -28,36 +28,40 @@ def get_internshipdata(db: Session = Depends(get_db)):
 @router.get("/{internship_id}")
 def get_internship(internship_id:int,db: Session = Depends(get_db)):
     internship = db.get(Internship,internship_id)
+    if not internship:
+        raise HTTPException(status_code=404,detail="Internship not found for the given id")
     return {"company":internship.company_name,"role":internship.role,"description":internship.description,"duration":internship.Duration}
 
 # to add completely a new internship to db
 @router.post("/add-internship")
 def add_internship(internship:InternData,db: Session = Depends(get_db)):
-    new_intern = Internship(company_name=internship.companyname,role=internship.role,description=internship.description,Duration=internship.duration)
+    new_intern = Internship(company_name=internship.company_name,role=internship.role,description=internship.description,Duration=internship.duration)
     db.add(new_intern)
     db.commit()
     db.refresh(new_intern)
-    return {"message":f"Internship:{new_intern.company_name} have been added successfully"}
+    return {"message":f"Internship: {new_intern.company_name} have been added successfully"}
 
 
 #to edit things in the internships
 @router.put("/edit-internship/{internship_id}")
 def edit_internship(internship:InternData,internship_id:int,db: Session = Depends(get_db)):
     curr_internship = db.query(Internship).filter(Internship.id==internship_id).first()
-    curr_internship.company_name = internship.companyname
+    curr_internship.company_name = internship.company_name
     curr_internship.description = internship.description
     curr_internship.role = internship.role
     curr_internship.Duration = internship.duration
     db.commit()
     db.refresh(curr_internship)
     
-    return {"message":f"Internship:{internship.companyname} has been updated successfully"}    
+    return {"message":f"Internship:{internship.company_name} has been updated successfully"}    
     
     
 #to delete a internship in the db
 @router.delete("/delete-internship/{internship_id}")
 def delete_internship(internship_id:int,db: Session = Depends(get_db)):
     internship = db.get(Internship,internship_id)
+    if not internship:
+        raise HTTPException(status_code=404,detail="Internship not found for the given id")
     db.delete(internship)
     db.commit()
     return {"message":f"Internship:{internship.company_name} have been deleted successfully"}
